+++ libs/core/src/main/java/org/opensearch/core/rest/RestStatus.java
@@ -38,6 +38,7 @@ import org.opensearch.core.common.io.stream.StreamOutput;
 
 import java.io.IOException;
 import java.util.HashMap;
+import java.util.Locale;
 import java.util.Map;
 
 import static java.util.Collections.unmodifiableMap;
@@ -432,6 +433,13 @@ public enum RestStatus {
      */
     EXPECTATION_FAILED(417),
     /**
+     * Any attempt to brew coffee with a teapot should result in the error code "418 I'm a teapot". The resulting
+     * entity body MAY be short and stout.
+     *
+     * @see <a href="https://www.rfc-editor.org/rfc/rfc2324#section-2.3.2">I'm a teapot!</a>
+     */
+    I_AM_A_TEAPOT(418),
+    /**
      * The request was directed at a server that is not able to produce a response. This can be sent by a server
      * that is not configured to produce responses for the combination of scheme and authority that are included
      * in the request URI.
@@ -558,5 +566,31 @@ public enum RestStatus {
      */
     public static RestStatus fromCode(int code) {
         return CODE_TO_STATUS.get(code);
+    }
+
+    public static boolean isValidRestCode(final int code) {
+        return fromCode(code) != null;
+    }
+
+    /**
+     * Get category class of a rest status code.
+     *
+     * @return String representing class category of the concrete rest status code
+     */
+    public String getClassName() {
+        return String.format(Locale.ROOT, "%dxx", status / 100);
+    }
+
+    /**
+     * Get category class of a rest status code.
+     *
+     * @return String representing class category of the concrete rest status code
+     */
+    public static String getClassName(final int code) {
+        if (isValidRestCode(code)) {
+            return fromCode(code).getClassName();
+        }
+
+        return null;
     }
 }