+++ server/src/main/java/org/opensearch/action/bulk/TransportBulkAction.java
@@ -79,6 +79,7 @@ import org.opensearch.index.VersionType;
 import org.opensearch.index.seqno.SequenceNumbers;
 import org.opensearch.core.index.shard.ShardId;
 import org.opensearch.indices.IndexClosedException;
+import org.opensearch.indices.IndicesService;
 import org.opensearch.indices.SystemIndices;
 import org.opensearch.ingest.IngestService;
 import org.opensearch.node.NodeClosedException;
@@ -129,6 +130,7 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
     private final IndexNameExpressionResolver indexNameExpressionResolver;
     private static final String DROPPED_ITEM_WITH_AUTO_GENERATED_ID = "auto-generated";
     private final IndexingPressureService indexingPressureService;
+    private final IndicesService indicesService;
     private final SystemIndices systemIndices;
 
     @Inject
@@ -143,6 +145,7 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
         IndexNameExpressionResolver indexNameExpressionResolver,
         AutoCreateIndex autoCreateIndex,
         IndexingPressureService indexingPressureService,
+        IndicesService indicesService,
         SystemIndices systemIndices
     ) {
         this(
@@ -156,6 +159,7 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
             indexNameExpressionResolver,
             autoCreateIndex,
             indexingPressureService,
+            indicesService,
             systemIndices,
             System::nanoTime
         );
@@ -172,6 +176,7 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
         IndexNameExpressionResolver indexNameExpressionResolver,
         AutoCreateIndex autoCreateIndex,
         IndexingPressureService indexingPressureService,
+        IndicesService indicesService,
         SystemIndices systemIndices,
         LongSupplier relativeTimeProvider
     ) {
@@ -187,6 +192,7 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
         this.client = client;
         this.indexNameExpressionResolver = indexNameExpressionResolver;
         this.indexingPressureService = indexingPressureService;
+        this.indicesService = indicesService;
         this.systemIndices = systemIndices;
         clusterService.addStateApplier(this.ingestForwarder);
     }
@@ -632,6 +638,8 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
                             if (bulkItemResponse.getResponse() != null) {
                                 bulkItemResponse.getResponse().setShardInfo(bulkShardResponse.getShardInfo());
                             }
+
+                            indicesService.incrementDocStatusCounter(bulkItemResponse.status());
                             responses.set(bulkItemResponse.getItemId(), bulkItemResponse);
                         }
                         if (counter.decrementAndGet() == 0) {
@@ -644,15 +652,15 @@ public class TransportBulkAction extends HandledTransportAction<BulkRequest, Bul
                         // create failures for all relevant requests
                         for (BulkItemRequest request : requests) {
                             final String indexName = concreteIndices.getConcreteIndex(request.index()).getName();
-                            DocWriteRequest<?> docWriteRequest = request.request();
-                            responses.set(
+                            final DocWriteRequest<?> docWriteRequest = request.request();
+                            final BulkItemResponse bulkItemResponse = new BulkItemResponse(
                                 request.id(),
-                                new BulkItemResponse(
-                                    request.id(),
-                                    docWriteRequest.opType(),
-                                    new BulkItemResponse.Failure(indexName, docWriteRequest.id(), e)
-                                )
+                                docWriteRequest.opType(),
+                                new BulkItemResponse.Failure(indexName, docWriteRequest.id(), e)
                             );
+
+                            indicesService.incrementDocStatusCounter(bulkItemResponse.status());
+                            responses.set(request.id(), bulkItemResponse);
                         }
                         if (counter.decrementAndGet() == 0) {
                             finishHim();