+++ server/src/test/java/org/opensearch/action/admin/cluster/node/stats/NodeStatsTests.java
@@ -32,6 +32,7 @@
 
 package org.opensearch.action.admin.cluster.node.stats;
 
+import org.opensearch.action.admin.indices.stats.CommonStats;
 import org.opensearch.cluster.node.DiscoveryNode;
 import org.opensearch.cluster.routing.WeightedRoutingStats;
 import org.opensearch.cluster.service.ClusterManagerThrottlingStats;
@@ -42,6 +43,8 @@ import org.opensearch.discovery.DiscoveryStats;
 import org.opensearch.cluster.coordination.PendingClusterStateStats;
 import org.opensearch.cluster.coordination.PublishClusterStateStats;
 import org.opensearch.http.HttpStats;
+import org.opensearch.index.shard.IndexingStats;
+import org.opensearch.indices.NodeIndicesStats;
 import org.opensearch.indices.breaker.AllCircuitBreakerStats;
 import org.opensearch.indices.breaker.CircuitBreakerStats;
 import org.opensearch.ingest.IngestStats;
@@ -65,6 +68,7 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
@@ -83,6 +87,31 @@ public class NodeStatsTests extends OpenSearchTestCase {
                 NodeStats deserializedNodeStats = new NodeStats(in);
                 assertEquals(nodeStats.getNode(), deserializedNodeStats.getNode());
                 assertEquals(nodeStats.getTimestamp(), deserializedNodeStats.getTimestamp());
+                if (nodeStats.getIndices() == null) {
+                    assertNull(deserializedNodeStats.getIndices());
+                } else {
+                    IndexingStats.Stats indexingStats = nodeStats.getIndices().getIndexing().getTotal();
+                    IndexingStats.Stats deserializedIndexingStats = deserializedNodeStats.getIndices().getIndexing().getTotal();
+
+                    assertEquals(indexingStats.getIndexCount(), deserializedIndexingStats.getIndexCount());
+                    assertEquals(indexingStats.getIndexTime(), deserializedIndexingStats.getIndexTime());
+                    assertEquals(indexingStats.getIndexCurrent(), deserializedIndexingStats.getIndexCurrent());
+                    assertEquals(indexingStats.getIndexFailedCount(), deserializedIndexingStats.getIndexFailedCount());
+                    assertEquals(indexingStats.getDeleteCount(), deserializedIndexingStats.getDeleteCount());
+                    assertEquals(indexingStats.getDeleteCurrent(), deserializedIndexingStats.getDeleteCurrent());
+                    assertEquals(indexingStats.getDeleteTime(), deserializedIndexingStats.getDeleteTime());
+                    assertEquals(indexingStats.getNoopUpdateCount(), deserializedIndexingStats.getNoopUpdateCount());
+                    assertEquals(indexingStats.isThrottled(), deserializedIndexingStats.isThrottled());
+                    assertEquals(indexingStats.getThrottleTime(), deserializedIndexingStats.getThrottleTime());
+
+                    Map<String, AtomicLong> indexingDocStatusCounter = indexingStats.getDocStatusStats().getDocStatusCounter();
+                    Map<String, AtomicLong> deserializedDocStatusCounter = deserializedIndexingStats.getDocStatusStats()
+                        .getDocStatusCounter();
+
+                    for (String key : indexingDocStatusCounter.keySet()) {
+                        assertEquals(indexingDocStatusCounter.get(key).longValue(), deserializedDocStatusCounter.get(key).longValue());
+                    }
+                }
                 if (nodeStats.getOs() == null) {
                     assertNull(deserializedNodeStats.getOs());
                 } else {
@@ -722,7 +751,7 @@ public class NodeStatsTests extends OpenSearchTestCase {
         return new NodeStats(
             node,
             randomNonNegativeLong(),
-            null,
+            createNodeIndicesStats(),
             osStats,
             processStats,
             jvmStats,
@@ -747,7 +776,48 @@ public class NodeStatsTests extends OpenSearchTestCase {
         );
     }
 
+    private static NodeIndicesStats createNodeIndicesStats() {
+        if (rarely()) {
+            return null;
+        }
+
+        Map<String, AtomicLong> docStatusCounter = new HashMap<>();
+
+        if (frequently()) {
+            int size = randomInt(5);
+
+            for (int i = 0; i < size; ++i) {
+                docStatusCounter.put(randomFrom("1xx", "2xx", "3xx", "4xx", "5xx"), new AtomicLong(randomNonNegativeLong()));
+            }
+        }
+
+        IndexingStats.Stats.DocStatusStats docStatusStats = new IndexingStats.Stats.DocStatusStats(docStatusCounter);
+        IndexingStats.Stats stats = new IndexingStats.Stats(
+            randomNonNegativeLong(),
+            randomNonNegativeLong(),
+            randomNonNegativeLong(),
+            randomNonNegativeLong(),
+            randomNonNegativeLong(),
+            randomNonNegativeLong(),
+            randomNonNegativeLong(),
+            randomNonNegativeLong(),
+            randomBoolean(),
+            randomNonNegativeLong(),
+            docStatusStats
+        );
+        IndexingStats indexingStats = new IndexingStats(stats);
+
+        return new NodeIndicesStats(
+            new CommonStats(null, null, indexingStats, null, null, null, null, null, null, null, null, null, null, null, null, null),
+            Collections.emptyMap()
+        );
+    }
+
     private OperationStats getPipelineStats(List<IngestStats.PipelineStat> pipelineStats, String id) {
-        return pipelineStats.stream().filter(p1 -> p1.getPipelineId().equals(id)).findFirst().map(p2 -> p2.getStats()).orElse(null);
+        return pipelineStats.stream()
+            .filter(p1 -> p1.getPipelineId().equals(id))
+            .findFirst()
+            .map(IngestStats.PipelineStat::getStats)
+            .orElse(null);
     }
 }